requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
exceptiongroup==1.3.0
    # via anyio
fastapi==0.116.1
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
python-crontab==3.3.0
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via -r requirements.in
starlette==0.47.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
